Add ticket cookie to identify the session
When multiple requests arrive concurrently with an expired access token,
the expected behavior is to have all succeed but instead, some fail
while the refresh process happens. That happens due to the key derived
from the last valid access token missing while other concurrent requests
search for it.

To prevent that race condition, we generate a static identifier instead
of one derived from a dynamic token. For that, we've borrowed the
ticketing strategy and code from [oauth-proxy] to generate a single
static identifier, which has its expiration refreshed every time the
access token change.

[oauth-proxy]: https://github.com/oauth2-proxy/oauth2-proxy