Fix usage of refresh-token expiration
Due to the usage of the same function that decodes an OIDC/ID Token for
decoding a refresh token, we couldn't correctly determine the expiration
date from the refresh token. Along with that, when issuing a new access
token from a previous refresh token, we used the wrong expiration time
and ended with a dead-lock since both refresh and access tokens had the
same expiration.

By creating a new function and renaming existing functions and configs,
I clarified the difference between what parses OIDC tokens and what is
used only to retrieve the value of an expiration claim from a token.
Besides, we're now giving a default deadline for every data that's
persisted, even when the refresh token has no expiration. This change
doesn't make continuous sessions impossible; it only establishes a
maximum interval between every refresh.